Working on Identity

+++ KNews.Core.Entities/Post.cs
@@ -36,7 +36,7 @@ namespace KNews.Core.Entities
         public string Title { get; set; }
         public string Content { get; set; }
         public string ShortContent { get; set; }
-        public int AuthorID { get; set; }
+        public long AuthorID { get; set; }
         public DateTime CreateDate { get; set; }
         public DateTime? UpdateDate { get; set; }
         public DateTime? DeleteDate { get; set; }
