+++ KNews.Core.Services/Posts/Entities/PostFull.cs
@@ -11,7 +11,7 @@ namespace KNews.Core.Services.Posts.Entities
         public string Title { get; set; }
         public string Content { get; set; }
         public string ShortContent { get; set; }
-        public int AuthorID { get; set; }        
+        public long AuthorID { get; set; }        
         public DateTime CreateDate { get; set; }
         public DateTime? UpdateDate { get; set; }
         public DateTime? DeleteDate { get; set; }
